After review with Zsuark and Paolo. Was working on Integration tests. Stopping to work on unit tests first.
