+++ config/test.json
@@ -1,6 +1,7 @@
 {
   // Persistence layer defaults
   "persistence": {
+    "default": "mongo",
     "mongo": {
       "database": "asyncvoter-test"
     }
