+++ services/persistence.js
@@ -6,20 +6,21 @@
 // Persistence layer modules are facades to the actual persistence objects & functions
 // A persistence object must contain: { setupVote, giveVote, getVotes, deleteAllData }
 
+const config = require('config')
 
 module.exports = (store, options) => {
 
   const storeDefined = store && typeof store == 'string' && store.length > 0
 
-  if (!storeDefined) console.warn('Not sure what persistence to use - defaulting back to redis')
+  const storeToUse = storeDefined ? store : (config.persistence.default || 'redis')
+  console.log("Using persistence store: " + storeToUse)
+  const moduleName = './persistence/' + storeToUse
 
-  const moduleName = './persistence/' + (storeDefined ? store : 'redis')
-
+  console.log("+++ Loading:", moduleName)
   const persistence = require(moduleName)(options)
 
   // NO destruction of non-test databases
   if (process.env.NODE_ENV != 'test') delete persistence.deleteAllData
-
   return persistence
 
 }
